# internal/app/db/sqlite/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package sqlc

import (
	"database/sql"
	"time"

	"github.com/alexgolang/ishare-task/internal/app/domain"
)

type Task struct {
	ID          string              `json:"id"`
	Title       string              `json:"title"`
	Description sql.NullString      `json:"description"`
	Status      domain.TaskStatus   `json:"status"`
	Priority    domain.TaskPriority `json:"priority"`
	CreatedAt   time.Time           `json:"created_at"`
	UpdatedAt   time.Time           `json:"updated_at"`
}